Working compilation of baseline 2 : 1 . Map palloc is added to xv6 malloc 2 . Tc1 passes

+++ starter-hw08/xmalloc.h
@@ -1,9 +1,12 @@
 #ifndef XMALLOC_H
 #define XMALLOC_H
 
+#include <pthread.h>
 #include <stddef.h>
+#include <stdio.h>
+#include <stdlib.h>
 
-void* xmalloc(size_t bytes);
+void* xmalloc(uint bytes);
 void  xfree(void* ptr);
 void* xrealloc(void* prev, size_t bytes);
